chore: re-organize code, move duplicate logic for tropo redirect to search service
